+++ packages/sandbox/CHANGELOG.md
@@ -1,5 +1,16 @@
 # sandbox
 
+## 3.0.0-beta.14
+
+### Minor Changes
+
+- Rebase from main
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/sandbox@2.0.0-beta.12
+
 ## 3.0.0-beta.13
 
 ### Patch Changes
@@ -127,6 +138,27 @@
 ### Major Changes
 
 - Support named sandboxes
+
+## 2.5.9
+
+### Patch Changes
+
+- Updated dependencies [[`9555162f33690dfa18530aeca93af05188ebd2ed`](https://github.com/vercel/sandbox/commit/9555162f33690dfa18530aeca93af05188ebd2ed)]:
+  - @vercel/sandbox@1.9.3
+
+## 2.5.8
+
+### Patch Changes
+
+- Updated dependencies [[`b91b9e49fb7d2c5a4b601c07c125e6e5a2c43441`](https://github.com/vercel/sandbox/commit/b91b9e49fb7d2c5a4b601c07c125e6e5a2c43441)]:
+  - @vercel/sandbox@1.9.2
+
+## 2.5.7
+
+### Patch Changes
+
+- Updated dependencies [[`cf13a34221c2b83c25c73d94929d05e0a697aecf`](https://github.com/vercel/sandbox/commit/cf13a34221c2b83c25c73d94929d05e0a697aecf), [`772989c59a3c27efa98153cdc54b6e35c1c15eae`](https://github.com/vercel/sandbox/commit/772989c59a3c27efa98153cdc54b6e35c1c15eae), [`184cd42d8d3b1ea1df354529cb6ba103a33e18d3`](https://github.com/vercel/sandbox/commit/184cd42d8d3b1ea1df354529cb6ba103a33e18d3), [`451c42efb94ab9c9dc330b4742071ac01008044d`](https://github.com/vercel/sandbox/commit/451c42efb94ab9c9dc330b4742071ac01008044d)]:
+  - @vercel/sandbox@1.9.1
 
 ## 2.5.6